requirements.txt
antspyx==0.5.4
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.54.1
gmsh==4.13.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
meshio==5.3.5
msvc_runtime==14.40.33807
numpy==2.0.1
numpy-stl==3.1.2
packaging==23.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pygments==2.18.0
pygmsh==7.1.17
pymeshfix==0.17.0
pymeshlab==2023.12.post2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-utils==3.9.0
pytz==2024.2
pyvista==0.44.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy==1.14.1
scooby==0.10.0
six==1.16.0
SQLAlchemy==2.0.27
statsmodels==0.14.4
tqdm==4.67.0
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
vtk==9.3.1
webcolors==24.11.1
Werkzeug==3.0.1
WTForms==3.1.2
